mockito-app/src/test/java/gt/com/archteam/mockitoapp/services/ExamenServiceImplTest.java
package gt.com.archteam.mockitoapp.services;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

import org.junit.jupiter.api.Test;

import gt.com.archteam.mockitoapp.models.Examen;
import gt.com.archteam.mockitoapp.repositories.ExamenRepository;
import gt.com.archteam.mockitoapp.repositories.ExamenRepositoryImpl;

class ExamenServiceImplTest {

    @Test
    void testFindExamenPorNombre() {
        ExamenRepository repository = new ExamenRepositoryImpl();
        ExamenService service = new ExamenServiceImpl(repository);
        Examen examen = service.findExamenPorNombre("Matematicas");
        
        assertNotNull(examen);
        assertEquals(5L, examen.getId());
        assertEquals("Matematicas", examen.getNombre());
    }
}
